feat!: http mock server (#83) (#128)
### Features
- Added support to remote HTTP interceptors! Now, it is possible to use
zimic in combination with a mock server to more complex use cases.
- The worker platform is now automatically inferred. Some exports were
changed by this simplification:
- Removed the exported error classes
`InvalidHttpInterceptorWorkerPlatform` and
`MismatchedHttpInterceptorWorkerPlatform`
- Added the exported error class `UnknownHttpInterceptorWorkerPlatform`,
when an unsupported platform is used.
- Added validation to service schemas, preventing the declaration of
body types for methods that do not support it by them:
  - GET is not expected to have a request body
  - HEAD is not expected to have a request or response body
  - OPTIONS is not expected to have a request body.

### Refactoring
- `tracker.requests()` array return is no longer appended after new
requests are intercepted. Now, the return is an immutable array and must
be called again to get the latest requests.

Closes #83.