feat(#zimic)!: internal worker management (#127) (#133)
### Features

- [#zimic] Simplified the interceptor factory `createHttpInterceptor` to
receive only a `type` and a `baseURL` as parameters. Creating, sharing,
starting and stopping workers are now performed internally. Therefore,
workers no longer have to be created and passed directly when creating
an interceptor, simplifying the API. 🎉
  With this change, the following exports were changed:
  - `createHttpInterceptorWorker` -> removed
  - `HttpInterceptorWorker` -> removed
  - `LocalHttpInterceptorWorker` -> removed
  - `RemoteHttpInterceptorWorker` -> removed
  - `HttpInterceptorWorkerOptions` -> removed
  - `LocalHttpInterceptorWorkerOptions` -> removed
  - `RemoteHttpInterceptorWorkerOptions` -> removed
  - `HttpInterceptorWorkerType` -> changed to `HttpInterceptorType`
- `HttpInterceptorWorkerPlatform` -> changed to
`HttpInterceptorPlatform`
- `UnknownHttpInterceptorWorkerPlatform` -> changed to
`UnknownHttpInterceptorPlatform`
- `NotStartedHttpInterceptorWorkerError` -> changed to
`NotStartedHttpInterceptorError`
  - `OtherHttpInterceptorWorkerRunningError` -> removed
- [#zimic] HTTP interceptors are now automatically cleared when they are
stopped. This is to prevent memory leaks. Previously, it was necessary
to manually clear each interceptors before stopping them.
- [#zimic] Added support to remote procedure calls over `wss://` when
`https://` is used as the base URL. This allows
  secure websocket clients to be used with the interceptors.

### Refactoring
- [#zimic] Split more tests into dedicated files to improve readability.
- [#zimic] Removed the server parameters `--life-cycle-timeout` and
`--rpc-timeout`. They were not used as of now and
were proving to be unnecessary to the current implementation. If the
need arises, they can be added back in the
  future.
- [#zimic] The errors `NoResponseDefinitionError` and
`UnknownHttpInterceptorTypeError` now extend `TypeError`, which is more
appropriate.

### Fixes
- [#zimic] Added missing test config files to `turbo.json` inputs.

### Examples

- [jest-jsdom, jest-node, vitest-jsdom, vitest-node, vitest-browser]
Updated the examples to use the simplified interceptor factory.

Closes #127.