feat: custom unhandled request strategies (#9) (#149)
### Features
- [#zimic] Added support to custom unhandled request strategies. Now,
the default configuration is `log: true`, meaning that requests that did
not match any handlers or interceptors are logged as a warning to the
console. When `type: 'local'`, unhandled requests are always bypassed
and might reach the real service if it is running. This is because we
have no way of "aborting" the request and returning a fetch error in the
local mode. However, when `type: 'remote'`, this is possible and
unhandled requests result in a fetch error to the client. The logging
configuration enables or disables only the warnings logged to the
console, not affecting the bypass vs reject behavior per interceptor
type.
- [#zimic] Improved the error messages to include `[#zimic]` at the
start.
- [#zimic] Included the restriction unhandled request strategy types to
exports:
  - `UnhandledRequestStrategy`
  - `HttpRequestHandlerRestriction`
  - `HttpRequestHandlerComputedRestriction`
  - `HttpRequestHandlerHeadersStaticRestriction`
  - `HttpRequestHandlerSearchParamsStaticRestriction`
  - `HttpRequestHandlerStaticRestriction`
  - `HttpRequestHandlerBodyStaticRestriction`

### Build
- [#zimic] Reverted `chalk` to version 4, which still supports CommonJS.
This is necessary to allow importing `chalk` from bundled applications
such as the Next.js and Playwright examples.

### Refactoring
- [#zimic] Freezed constant objects to avoid editions.

Example log for an unhandled request:


![image](https://github.com/diego-aquino/zimic/assets/58959382/971cc93f-e35d-4ff4-afce-45480f98ff8e)

Closes #9.