fix(#zimic): partial restriction types (#216)
### Fixes
- [#zimic] Updated the restriction types to allow partial matches.
Previously, a static restriction would have the same type as the header,
search params or body. However, the restriction matches are partial by
default, so it should be possible to specify partial headers, search
params and body restrictions.
- [#zimic] Fixed the types `StrictRequest` and `StrictResponse` to
correctly type the methods `.json() and `.formData()`, considering if
the typed body is a JSON value or a form data.
- [#zimic] Removed the type `ReadableStream` from `HttpBody`, no longer
allowing request and response body schemas to have a readable stream as
body type. This type was incorrect because `ReadableStream` objects are
not supported. If you are working with streams, consider transforming
them to an `ArrayBuffer` or `Blob`, since these are fully supported.

### Tests
- [#zimic] Added more checks to verify the types used in raw requests
and responses.
- [#zimic] Added more tests to verify `ArrayBuffer`'s can be used as
request and response body types.

Closes #212.