feat(#zimic): opt-in request saving (#210) (#224)
### Features
- [#zimic] Added the new interceptor boolean option `saveRequests`,
which is `false` by default. This option indicates whether request
handler should save their intercepted requests in memory and make them
accessible through `handler.requests()`. This is an opt-in feature
because it would cause a memory leak in long-running mock servers, if
not accompanied by `interceptor.clear()` calls or disposal of the
handlers by garbage collection.

Closes #210, closes #28.