fix(#zimic): skip remote clearing if web socket client is not running (#287)
### Fixes
- [#zimic] Added a check to only apply clearing operations to a remote
interceptor server if the web socket client is running. Previously,
sending Ctrl + C to the `load.ts` scripts in the Playwright and Next.js
App Router examples would cause an error. In those cases, the Ctrl + C
would already be handled by the interceptor server, which would close
all socket connections and prepare to stop. Therefore, it is expected
that the web socket client may be already closed when
`interceptor.stop()` is called and this should not cause an error.

### Refactoring
- [#zimic] Removed unnecessary `crypto` attributes in some classes.