test(zimic-test-client): use fixed github openapi commit (#378)
### Testing
- It appears that some change was merged yesterday to [GitHub's OpenAPI
v3.1
schema](https://github.com/github/rest-api-description/blob/main/descriptions-next/api.github.com/api.github.com.json).
Current, the file on `main` is broken and uses a ref that is not
declared. This is causing our end-to-end typegen tests to fail. While
they fix the problem, I fixed the commit we use to
https://github.com/github/rest-api-description/tree/8bec6e8fe8562e79a6034437e4e097e10df9cd2c,
which contains the latest working spec. We should revert back to `main`
after a fix is pushed to https://github.com/github/rest-api-description.