docs(wiki): move `interceptor.clear()` to `beforeEach()` (#516)
After the new `checkTimes()` API, introduced in
https://github.com/zimicjs/zimic/issues/398, the initial examples showed
`interceptor.clear()` being called after `interceptor.checkTimes()`.
This is not optimal because `checkTimes()` throws an error if some
handler has intercepted fewer or more than the defined requests, causing
the `clear()` to not run. Because of that, any subsequent test would
also fail on the `checkTimes()` call. This PR changes the examples to
use `interceptor.clear()` in a `beforeEach` hook, leaving only
`interceptor.checkTimes()` in the `afterEach`, making sure one does not
block the other from running.

Another strategy is to use a `try..finally` and use a single
`afterEach`.

```ts
afterEach(() => {
  try {
    interceptor.checkTimes()
  } finally {
    interceptor.clear()
  }
})
```

We're opting into `beforeEach` + `afterEach` to keep the test hooks flat
(minimize nesting) and keep the examples readable and simple.