feat(#zimic-http): include `content-type` headers in all typegen bodies (#599)
### Features
- [#zimic-http] Changed the OpenAPI typegen implementation to include
the header `content-type` in all generated request and responses.
Previously, the header was added only when more than one content type
existed in the OpenAPI schema. It is more accurate to always include
this header, since it will always be defined in the schema when a body
is specified. Also, `@zimic/fetch` requires that all usable headers are
defined in the schema to correctly encode request bodies.