fix(#zimic-fetch): merge headers and search params with defaults (#606)
### Fixes
- [fetch] When using default headers and/or search params in a fetch
client, headers or search params declared in a request would not merge
with the defaults, but complete override them in the request. This was
not expected, because headers and search params should be merged with
the defaults in a request, preferring the data of the request. This PR
fixes this issue.

```ts
const fetch = createFetch<MySchema>({
  baseURL: 'http://localhost:3000',
  headers: { 'accept-language': 'en' },
});

const response = await fetch('/users', {
  method: 'GET',
  headers: { 'user-agent': 'client' }
})
// The request will have { 'accept-language': 'en', 'user-agent': 'client' } as headers.