feat(interceptor): show warning when safe number of saved requests is exceeded (#349) (#659)
### Features
- [interceptor] Changed the HTTP interceptor option from `saveRequests`
to `requestSaving`, which is now an object with the properties `enabled`
and `safeLimit`.

  ```diff
  const interceptor = createHttpInterceptor<Schema>({
    type: 'local',
    baseURL: 'http://localhost:3000',
  - saveRequests: true,
  + requestSaving: { enabled: true },
  })
  ```

- [interceptor] Added the HTTP interceptor option
`requestSaving.safeLimit` to indicate how many requests are safe to save
in memory. Saving more than such limit will cause warnings in the
console, which let users know they should use `interceptor.clear()`,
increase the limit, or set `requestSaving.enabled` to false.

  ```ts
  const interceptor = createHttpInterceptor<Schema>({
    type: 'local',
    baseURL: 'http://localhost:3000',
    requestSaving: { enabled: true, safeLimit: 1000 },
  })
  ```

Closes #349.