fix(fetch): empty and non-empty request type inference (#699)
### Fixes
- [fetch] The fetch client was not correctly typing the request
parameters if the schema had an empty or non-empty `request` key, such
as shown below. In this case, `headers` and `searchParams` were inferred
to `unknown` and any value was accepted in the `fetch` calls. The types
were working only when no `request` existed. This PR fixes this issue.

```ts
type Schema = HttpSchema<{
  '/users': {
    GET: {
      request: {}; // empty request causing 
      response: { 200: { body: User[] } };
    };
  };
}>;


const fetch = createFetch<Schema>({ baseURL });

await fetch('/users', {
  method: 'GET',
  headers: { 'my-custom-headers: 'example' }, // should be a type error because the schema has no headers
  searchParams: { option: '1' } // should be a type error because the schema has no search params
}),
```

### Performance
- [fetch] Optimized the `Fetch` type, which was previously an
intersection and now is a interface extending `FetchClient`. This was
necessary after the fixes in the request type inference.