feat(http)!: json content type inference (#750) (#759)
### Features
- [http] Added support to infer the `content-type` header as
`application/json` when the body is a JSON object. If a content type is
declared, it is used instead, keeping backward compatibility.
- [http] Improved the serialization types of headers, search params,
path params, form data, now considering unserializable types as `string`
instead of `never`. This is more aligned with the real-word behavior.
For example, trying to add a Map as a search param value saves it as a
string.

### Refactoring
- [http] Moved the strict schema of `HttpHeaders`, `HttpSearchParams`,
and `HttpFormData` to an attribute instead of a generic parameter. This
is an optimization to keep the strict schema internal.

Closes #750.