fix(interceptor): consider null as empty response body (#1033)
There was a missing `declaration.body === null` condition here:


https://github.com/zimicjs/zimic/blob/93bf68204ef30f40576061e02b79b20ae4f07942/packages/zimic-interceptor/src/http/interceptorWorker/HttpInterceptorWorker.ts#L211

It caused response declarations with `body: null` to be interpreted as
JSON, thus returning the response body as a string (`'null'`), instead
of an empty body as the Fetch API. This is incorrect and our test suite
was not converting this scenario.

This PR adds the missing condition and increments our tests to correctly
check that `body: null` should lead to an empty response body.