fix(http): type `json()` and `formData()` as never if body is empty (#1065)
`request.json()` and `response.json()` throw an error if the body is
empty.

```bash
SyntaxError: Unexpected end of JSON input
    at JSON.parse (<anonymous>)
    at parseJSONFromBytes (node:internal/deps/undici/undici:5738:19)
    at successSteps (node:internal/deps/undici/undici:5719:27)
    at consumeBody (node:internal/deps/undici/undici:5725:9)
    at Response.json (node:internal/deps/undici/undici:5665:18)
```

`request.formData()` and `response.formData()` act similarly.

```bash
TypeError: Failed to parse body as FormData.
    at node:internal/deps/undici/undici:5675:27
    at successSteps (node:internal/deps/undici/undici:5719:27)
    at consumeBody (node:internal/deps/undici/undici:5725:9)
    at Response.formData (node:internal/deps/undici/undici:5668:18)
```

However, the types `HttpRequest` and `HttpResponse` in `@zimic/http`
typed the return of `json()` and `formData()` as `null`. The more
appropriate type is `never`, as these methods won't return when the body
is empty.