fix(interceptor): ignore abort errors when clearing a remote worker (#1120)
When stopping a remote interceptor at the same time as the interceptor
server, the interceptor worker can throw a `WebSocketMessageAbortError`.

```
.../node_modules/@zimic/interceptor/src/webSocket/WebSocketHandler.ts:394
    const abortError = new WebSocketMessageAbortError();
                       ^

WebSocketMessageAbortError: Web socket message was aborted.
    at WebSocketClient.abortSocketMessages (.../node_modules/@zimic/interceptor/src/webSocket/WebSocketHandler.ts:394:24)
    at WebSocketClient.removeSocket (.../node_modules/@zimic/interceptor/src/webSocket/WebSocketHandler.ts:197:10)
    at _WebSocket.handleSocketClose (.../node_modules/@zimic/interceptor/src/webSocket/WebSocketHandler.ts:78:12)
    at callListener (.../node_modules/isomorphic-ws/node_modules/ws/lib/event-target.js:290:14)
    at _WebSocket.onClose (.../node_modules/isomorphic-ws/node_modules/ws/lib/event-target.js:220:9)
    at _WebSocket.emit (node:events:508:28)
    at _WebSocket.emitClose (.../node_modules/isomorphic-ws/node_modules/ws/lib/websocket.js:272:10)
    at Socket.socketOnClose (.../node_modules/isomorphic-ws/node_modules/ws/lib/websocket.js:1341:15)
    at Socket.emit (node:events:508:28)
    at TCP.<anonymous> (node:net:346:12)

Node.js v24.11.1
```

This is because remote interceptors send a reset message to the
interceptor server when they are stopped, to indicate that the
interceptor should no longer be considered to match requests. If the
interceptor server is shutting down during this process, the reset
message may be aborted, which is expected and should not bubble up to
the client as an unhandled error.