fix(interceptor): abort websocket requests of removed handlers (#1132)
### Fixes

- [interceptor] Improved the logic for canceling WebSocket requests when
interceptor workers or interceptor servers are stopped. Previously,
there could be a rare racing condition in the abort logic, causing some
requests to not be aborted and triggering an unexpected
`WebSocketMessageTimeoutError`.

On top of removing relevant listeners before stopping the WebSocket
client and server, I also improved the abortion implementation to only
cancel requests of removed handlers, preserving requests of handlers
that are still being used.


https://github.com/zimicjs/zimic/blob/311fcf275496535dd86667faaccb93f73278a352/packages/zimic-interceptor/src/server/InterceptorServer.ts#L202-L213

### Refactoring

- [interceptor] Improved WebSocket event methods
- [root] Standardized `.gitignore` and ESLint ignores for consistency
- [root] Removed an unnecessary postinstall script in the root
package.json